+++ src/File.h
@@ -12,12 +12,12 @@ public:
 	virtual ~File(void);
 	virtual void OpenOld(const std::string &fileName);
 	virtual void OpenNew(const std::string &fileName);
-	virtual long Length(void);
-	virtual void SetPos(long pos, bool truncate);
-	virtual long GetPos(void);
-	virtual void ReadBytes(long numBytes, void *buffer);
-	virtual void ReadUnicode(long *len, wchar_t **text);
-	virtual void WriteBytes(long numBytes, void *buffer);
+	virtual int32_t Length(void);
+	virtual void SetPos(int32_t pos, bool truncate);
+	virtual int32_t GetPos(void);
+	virtual void ReadBytes(int32_t numBytes, void *buffer);
+	virtual void ReadUnicode(int32_t *len, wchar_t **text);
+	virtual void WriteBytes(int32_t numBytes, void *buffer);
 	virtual void Close(bool truncate);
 private:
 	bool m_error;
