+++ src/TMTParser.h
@@ -6,9 +6,9 @@
 
 class TMTParser {
 public:
-	virtual void Parse(bool *changedSrc, long *errPos, long *errLen, wchar_t error[]);
+	virtual void Parse(bool *changedSrc, int32_t *errPos, int32_t *errLen, wchar_t error[]);
 #if _DEBUG
-	virtual void RemoveAltCodePath(bool *changedSrc, long *errPos, long *errLen, wchar_t error[]);
+	virtual void RemoveAltCodePath(bool *changedSrc, int32_t *errPos, int32_t *errLen, wchar_t error[]);
 #endif
 	virtual void InitTMTParser(TextBuffer *talkText, TrueTypeFont *font, TrueTypeGlyph *glyph, bool legacyCompile, short generators, TTGenerator *gen[]);
 	virtual void TermTMTParser(void);
@@ -17,10 +17,10 @@ public:
 };
 
 TMTParser *NewTMTSourceParser(void);
-bool TMTCompile(TextBuffer *talkText, TrueTypeFont *font, TrueTypeGlyph *glyph, long glyphIndex, TextBuffer *glyfText, bool legacyCompile, long *errPos, long *errLen, wchar_t errMsg[]); // returns true if compilation completed successfully
+bool TMTCompile(TextBuffer *talkText, TrueTypeFont *font, TrueTypeGlyph *glyph, int32_t glyphIndex, TextBuffer *glyfText, bool legacyCompile, int32_t *errPos, int32_t *errLen, wchar_t errMsg[]); // returns true if compilation completed successfully
 
 #if _DEBUG
-bool TMTRemoveAltCodePath(TextBuffer *talkText, TrueTypeFont *font, TrueTypeGlyph *glyph, long *errPos, long *errLen, wchar_t errMsg[]);
+bool TMTRemoveAltCodePath(TextBuffer *talkText, TrueTypeFont *font, TrueTypeGlyph *glyph, int32_t *errPos, int32_t *errLen, wchar_t errMsg[]);
 #endif
 
 #endif // TMTParser_dot_h
